# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
  username?: string;
}

export type ReorderInputOrderItem = {
  id: number;
  displayOrder: number;
};

/**
 * New display order for a set of sibling records.
 */
export interface ReorderInput {
  order: ReorderInputOrderItem[];
}

export type ResourceType = typeof ResourceType[keyof typeof ResourceType];


export const ResourceType = {
  Lecture_notes: 'Lecture notes',
  Previous_year_paper: 'Previous year paper',
  Lab_manual: 'Lab manual',
  Assignment: 'Assignment',
  Reference: 'Reference',
} as const;

export type SubmissionStatus = typeof SubmissionStatus[keyof typeof SubmissionStatus];


export const SubmissionStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface Branch {
  id: number;
  name: string;
  shortName: string;
  description: string;
  displayOrder: number;
  isActive: boolean;
  /** Derived count of subjects under this branch. */
  subjectCount: number;
  /** Derived count of resources under this branch. */
  resourceCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface BranchInput {
  name: string;
  shortName: string;
  description?: string;
  displayOrder?: number;
  isActive?: boolean;
}

export interface BranchUpdate {
  name?: string;
  shortName?: string;
  description?: string;
  displayOrder?: number;
  isActive?: boolean;
}

export interface Year {
  id: number;
  branchId: number;
  name: string;
  displayOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface YearInput {
  branchId: number;
  name: string;
  displayOrder?: number;
}

export interface YearUpdate {
  branchId?: number;
  name?: string;
  displayOrder?: number;
}

export interface Semester {
  id: number;
  yearId: number;
  name: string;
  displayOrder: number;
  createdAt: string;
  updatedAt: string;
}

export interface SemesterInput {
  yearId: number;
  name: string;
  displayOrder?: number;
}

export interface SemesterUpdate {
  yearId?: number;
  name?: string;
  displayOrder?: number;
}

export interface Subject {
  id: number;
  semesterId: number;
  name: string;
  description: string;
  createdAt: string;
  updatedAt: string;
}

export interface SubjectInput {
  semesterId: number;
  name: string;
  description?: string;
}

export interface SubjectUpdate {
  semesterId?: number;
  name?: string;
  description?: string;
}

export interface Resource {
  id: number;
  subjectId: number;
  title: string;
  description: string;
  resourceType: ResourceType;
  googleDriveUrl: string;
  isNew: boolean;
  isFeatured: boolean;
  isVerified: boolean;
  verifiedAt?: string | null;
  verifiedBy?: string | null;
  createdAt: string;
  updatedAt: string;
  /** Derived from the related subject for display convenience. */
  subjectName?: string;
  semesterName?: string;
  yearName?: string;
  branchId?: number;
  branchName?: string;
}

export interface ResourceInput {
  subjectId: number;
  title: string;
  description?: string;
  resourceType: ResourceType;
  googleDriveUrl: string;
  isNew?: boolean;
  isFeatured?: boolean;
  isVerified?: boolean;
}

export interface ResourceUpdate {
  subjectId?: number;
  title?: string;
  description?: string;
  resourceType?: ResourceType;
  googleDriveUrl?: string;
  isNew?: boolean;
  isFeatured?: boolean;
  isVerified?: boolean;
}

export interface Submission {
  id: number;
  branchId?: number | null;
  yearId?: number | null;
  semesterId?: number | null;
  subjectId?: number | null;
  resourceType: ResourceType;
  title: string;
  description: string;
  googleDriveUrl: string;
  studentName: string;
  studentEmail: string;
  status: SubmissionStatus;
  adminNote?: string | null;
  rejectionReason?: string | null;
  submittedAt: string;
  reviewedAt?: string | null;
  reviewedBy?: string | null;
}

export interface SubmissionInput {
  branchId?: number;
  yearId?: number;
  semesterId?: number;
  subjectId?: number;
  resourceType: ResourceType;
  title: string;
  description?: string;
  googleDriveUrl: string;
  studentName: string;
  studentEmail: string;
}

export interface SubmissionUpdate {
  status?: SubmissionStatus;
  adminNote?: string;
  rejectionReason?: string;
}

export interface ApproveSubmissionInput {
  adminNote?: string;
  isFeatured?: boolean;
  isVerified?: boolean;
}

export interface RejectSubmissionInput {
  rejectionReason?: string;
}

/**
 * Include inactive branches. Only honored for authenticated admin sessions.
 */
export type IncludeInactiveParameter = boolean;

export type ListBranchesParams = {
/**
 * Include inactive branches. Only honored for authenticated admin sessions.
 */
includeInactive?: IncludeInactiveParameter;
};

export type ListYearsParams = {
branchId?: number;
};

export type ListSemestersParams = {
yearId?: number;
};

export type ListSubjectsParams = {
semesterId?: number;
};

export type ListResourcesParams = {
branchId?: number;
yearId?: number;
semesterId?: number;
subjectId?: number;
resourceType?: ResourceType;
isVerified?: boolean;
isNew?: boolean;
isFeatured?: boolean;
/**
 * Free-text search across title and description
 */
query?: string;
};

export type ListSubmissionsParams = {
status?: SubmissionStatus;
};