# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Admin login
      description: Authenticates an admin and starts a session (session cookie is set on the response).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthStatus"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List branches
      parameters:
        - $ref: "#/components/parameters/IncludeInactive"
      responses:
        "200":
          description: Branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create a branch
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/reorder:
    post:
      operationId: reorderBranches
      tags: [branches]
      summary: Reorder branches
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderInput"
      responses:
        "200":
          description: Reordered branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
  /branches/{id}:
    get:
      operationId: getBranch
      tags: [branches]
      summary: Get a branch by id
      parameters:
        - $ref: "#/components/parameters/BranchId"
      responses:
        "200":
          description: Branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update a branch
      security: [{ adminSession: [] }]
      parameters:
        - $ref: "#/components/parameters/BranchId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: Updated branch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete a branch
      security: [{ adminSession: [] }]
      parameters:
        - $ref: "#/components/parameters/BranchId"
      responses:
        "204":
          description: Deleted
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List resources
      description: Supports filtering by catalog position, type, and flags, plus free-text search over title/description.
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: yearId
          in: query
          schema:
            type: integer
        - name: semesterId
          in: query
          schema:
            type: integer
        - name: subjectId
          in: query
          schema:
            type: integer
        - name: resourceType
          in: query
          schema:
            $ref: "#/components/schemas/ResourceType"
        - name: isVerified
          in: query
          schema:
            type: boolean
        - name: isNew
          in: query
          schema:
            type: boolean
        - name: isFeatured
          in: query
          schema:
            type: boolean
        - name: query
          in: query
          description: Free-text search across title and description
          schema:
            type: string
      responses:
        "200":
          description: Resources
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resource"
    post:
      operationId: createResource
      tags: [resources]
      summary: Create a resource
      security: [{ adminSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{id}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Get a resource by id
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateResource
      tags: [resources]
      summary: Update a resource
      security: [{ adminSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResourceUpdate"
      responses:
        "200":
          description: Updated resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Delete a resource
      security: [{ adminSession: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: nexora_session
      description: Signed, httpOnly admin session cookie set by POST /auth/login.
  parameters:
    BranchId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
    IncludeInactive:
      name: includeInactive
      in: query
      description: Include inactive branches. Only honored for authenticated admin sessions.
      schema:
        type: boolean
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    LoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    AuthStatus:
      type: object
      properties:
        authenticated:
          type: boolean
        username:
          type: string
      required:
        - authenticated
    ReorderInput:
      type: object
      description: New display order for a set of sibling records.
      properties:
        order:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              displayOrder:
                type: integer
            required:
              - id
              - displayOrder
      required:
        - order
    ResourceType:
      type: string
      enum:
        - "Lecture notes"
        - "Previous year paper"
        - "Lab manual"
        - "Assignment"
        - "Reference"
...
    Branch:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        description:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
        subjectCount:
          type: integer
          description: Derived count of subjects under this branch.
        resourceCount:
          type: integer
          description: Derived count of resources under this branch.
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - shortName
        - description
        - displayOrder
        - isActive
        - subjectCount
        - resourceCount
        - createdAt
        - updatedAt
    BranchInput:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        description:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
      required:
        - name
        - shortName
    BranchUpdate:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        description:
          type: string
        displayOrder:
          type: integer
        isActive:
          type: boolean
...
    Resource:
      type: object
      properties:
        id:
          type: integer
        subjectId:
          type: integer
        title:
          type: string
        description:
          type: string
        resourceType:
          $ref: "#/components/schemas/ResourceType"
        googleDriveUrl:
          type: string
        isNew:
          type: boolean
        isFeatured:
          type: boolean
        isVerified:
          type: boolean
        verifiedAt:
          type: string
          format: date-time
          nullable: true
        verifiedBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        subjectName:
          type: string
          description: Derived from the related subject for display convenience.
        semesterName:
          type: string
        yearName:
          type: string
        branchId:
          type: integer
        branchName:
          type: string
      required:
        - id
        - subjectId
        - title
        - description
        - resourceType
        - googleDriveUrl
        - isNew
        - isFeatured
        - isVerified
        - createdAt
        - updatedAt
    ResourceInput:
      type: object
      properties:
        subjectId:
          type: integer
        title:
          type: string
        description:
          type: string
        resourceType:
          $ref: "#/components/schemas/ResourceType"
        googleDriveUrl:
          type: string
        isNew:
          type: boolean
        isFeatured:
          type: boolean
        isVerified:
          type: boolean
      required:
        - subjectId
        - title
        - resourceType
        - googleDriveUrl
    ResourceUpdate:
      type: object
      properties:
        subjectId:
          type: integer
        title:
          type: string
        description:
          type: string
        resourceType:
          $ref: "#/components/schemas/ResourceType"
        googleDriveUrl:
          type: string
        isNew:
          type: boolean
        isFeatured:
          type: boolean
        isVerified:
          type: boolean
...