# lib/api-zod/src/generated/types/resource.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceType } from './resourceType';

export interface Resource {
  id: number;
  subjectId: number;
  title: string;
  description: string;
  resourceType: ResourceType;
  googleDriveUrl: string;
  isNew: boolean;
  isFeatured: boolean;
  isVerified: boolean;
  verifiedAt?: string | null;
  verifiedBy?: string | null;
  createdAt: string;
  updatedAt: string;
  /** Derived from the related subject for display convenience. */
  subjectName?: string;
  semesterName?: string;
  yearName?: string;
  branchId?: number;
  branchName?: string;
}