# lib/api-zod/src/generated/types/resourceInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ResourceType } from './resourceType';

export interface ResourceInput {
  subjectId: number;
  title: string;
  description?: string;
  resourceType: ResourceType;
  googleDriveUrl: string;
  isNew?: boolean;
  isFeatured?: boolean;
  isVerified?: boolean;
}