# lib/api-zod/src/generated/types/resourceType.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type ResourceType = typeof ResourceType[keyof typeof ResourceType];


export const ResourceType = {
  Lecture_notes: 'Lecture notes',
  Previous_year_paper: 'Previous year paper',
  Lab_manual: 'Lab manual',
  Assignment: 'Assignment',
  Reference: 'Reference',
} as const;